+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Added the React `initialValue` prop and forwarded `value` and `initialValue` directly as React 19 custom-element properties.
+
 ### Fixed
 
 - Corrected the React build external and UMD global mapping from `jb-input-react` to `jb-input/react`.